Consolidate test store helpers and drop unused harmonia deps
The daemon-protocol-only decision (PLAN.md Open Question 11) makes
three Phase 0 dependencies dead weight:

  - harmonia-store-path-info: UnkeyedValidPathInfo reaches us through
    the harmonia-store-remote re-export
  - harmonia-store-nar-info: not needed until the Phase 4 substituter
    builds narinfo responses
  - harmonia-utils-base-encoding: nothing needs raw base32 encoding

Keeping unused git deps costs build time on every CI run and suggests
code paths that do not exist; re-adding a pinned line later is free.

Also move the daemon-availability probe and store-path/nix-oracle
helpers from individual test files into tests/support/store.rs:
the upcoming pipeline and serve tests need the same
"skip when no daemon is reachable" gating, and duplicating the
probe per test file is how skips drift apart.
